llama4-document-processor/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langgraph==0.2.50
langchain==0.3.10
langchain-community==0.3.10
ollama==0.1.7
boto3==1.34.0
python-docx==1.1.0
PyPDF2==3.0.1
pydantic>=2.7.4
python-dotenv==1.0.0